+++ R/insert_index_sheet.R
@@ -2,6 +2,11 @@
 #'
 #' @description Function which generates an index sheet inside an openxlsx
 #'   workbook.
+#' @details The logo can be customized via the `statR_logo` global option. This
+#'   should point to the path of the logo file. Other options such as the image
+#'   size in the final .xlsx can either be changed via the options
+#'   `statR_logo_width` and `statR_logo_height`, or set along with contact
+#'   information in a custom profile.
 #' @inheritParams insert_worksheet
 #' @param openinghours statzh or a character string or vector with opening hours
 #' @param auftrag_id Order ID
@@ -15,12 +20,10 @@ insert_index_sheet <- function(
   # Initialize new worksheet as index sheet ------
   openxlsx::addWorksheet(wb, sheetname)
 
-
   # Insert logo ----------
   insert_worksheet_image(
-    wb = wb, sheetname = sheetname, image = inputHelperLogoPath(logo), startrow = 1,
-    startcol = 1, width = 2.145, height = 0.7865)
-
+    wb = wb, sheetname = sheetname, image = inputHelperLogoPath(logo),
+    startrow = 1, startcol = 1)
 
   # Insert contact info, title, metadata, and sources into worksheet --------
   ### Contact information
@@ -43,12 +46,14 @@ insert_index_sheet <- function(
 
   ### Add Headerline
   openxlsx::addStyle(
-    wb, sheetname, style_headerline(), namedRegionLastRow(wb, sheetname, "homepage") + 1, 1:20,
+    wb, sheetname, style_headerline(),
+    namedRegionLastRow(wb, sheetname, "homepage") + 1, 1:20,
     gridExpand = TRUE, stack = TRUE)
 
   ### Request information
   openxlsx::writeData(
-    wb, sheetname, c(inputHelperDateCreated(), inputHelperOrderNumber(auftrag_id)),
+    wb, sheetname, c(inputHelperDateCreated(),
+                     inputHelperOrderNumber(auftrag_id)),
     startCol = 15, startRow = namedRegionLastRow(wb, sheetname, "homepage") + 3,
     name = paste(sheetname, "info", sep = "_"))
 
@@ -58,7 +63,8 @@ insert_index_sheet <- function(
     name = paste(sheetname, "title", sep = "_"))
 
   openxlsx::addStyle(
-    wb, sheetname, style_maintitle(), namedRegionLastRow(wb, sheetname, "title"), 3)
+    wb, sheetname, style_maintitle(),
+    namedRegionLastRow(wb, sheetname, "title"), 3)
 
   ### Source
   openxlsx::writeData(
@@ -68,19 +74,20 @@ insert_index_sheet <- function(
 
   ### Table of content caption
   openxlsx::writeData(
-    wb, sheetname, getOption("statR_toc_title"), 3, namedRegionLastRow(wb, sheetname, "source") + 3,
+    wb, sheetname, getOption("statR_toc_title"), 3,
+    namedRegionLastRow(wb, sheetname, "source") + 3,
     name = paste(sheetname, "toc", sep = "_"))
 
   openxlsx::addStyle(
     wb, sheetname, subtitleStyle(), namedRegionLastRow(wb, sheetname, "toc"), 3)
 
-  purrr::walk(
-    namedRegionFirstRow(wb, sheetname, "title"):namedRegionFirstRow(wb, sheetname, "toc"),
-    ~openxlsx::mergeCells(wb, sheetname, cols = 3:13, rows = .x))
-
-  purrr::walk(
-    namedRegionFirstRow(wb, sheetname, "title"):namedRegionFirstRow(wb, sheetname, "toc"),
-    ~openxlsx::addStyle(wb, sheetname, style_wrap(), cols = 3:13, rows = .x, stack = TRUE))
+  # purrr::walk(
+  #   namedRegionRowExtent(wb, sheetname, c("title", "source", "toc")),
+  #   ~openxlsx::mergeCells(wb, sheetname, cols = 3:13, rows = .x))
+  #
+  # purrr::walk(
+  #   namedRegionRowExtent(wb, sheetname, c("title", "source", "toc")),
+  #   ~openxlsx::addStyle(wb, sheetname, style_wrap(), cols = 3:13, rows = .x, stack = TRUE))
 
   # Format ---------
   ### Set column width of first column to 1